Prevent kernel to be updated by dist-upgrade
Tipically when a environment is provisioned it usually generates
specific drives for the kernel that is running, and sometimes whith a
apt-get upgrade or apt-get dist-upgrade happens it replaces the kernel
breaking the modules for the kernel that were made for that particular
kernel at the time of provision.

This commit adds a lock to the kernel running in the box at the time of
the provision.